requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
braintree==4.17.1
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-paypal==2.0
django-phonenumber-field==7.0.1
django-taggit==3.1.0
djangorestframework==3.14.0
fonttools==4.38.0
html5lib==1.1
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
phonenumbers==8.13.2
phonenumberslite==8.13.2
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyphen==0.13.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
weasyprint==57.2
webencodings==0.5.1
whitenoise==6.2.0
zopfli==0.2.2